+++ Base/Arpeggiator/Interface/Arpeggiator.h
@@ -49,7 +49,7 @@ public:
    CB_SIGNAL_IF(RangeTypeChanged, RangeType);
    CB_SIGNAL_IF(RangeChanged, int);
    CB_SIGNAL_IF(GateFillChanged, float);
-   CB_SIGNAL_IF(StepLengthChanged, int);
+   CB_SIGNAL_IF(StepLengthChanged, float);
    CB_SIGNAL_IF(AlgorithmChanged, Algorithm);
    CB_SIGNAL_IF(HoldNotesChanged, bool);
    CB_SIGNAL_IF(FeedModeChanged, FeedMode);
